fix: properly sets the `DELAY` default
Using `??` doesn't work when `localStorage.getItem(delayKey)` is used as i will by default return and empty string (`""`) which will set `DELAY = ""` instead of using the default value used. A better option is to use the OR operator (`||`) as it will return the right-hand side value if the left-hand side value is either `null`, `undefined`, or `""`.