feat: improved queries for selecting elements and added a functionality to remove reactions of chat messages from users
Some of the query strings miss some of the HTML elements, such as
partner's chat where they replied to one of the user's messages. These
new query strings capture the elements properly. Removal of user's
reactions to chat messages was added as a feature and it is enabled by
default.